# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prioritize API",
    "description": "REST API of Prioritize, a resource, project and IoT orchestration platform. All endpoints live under /api/v1. From 1.0.0 on this REST surface is the stable public contract (SemVer); the Vaadin admin GUI and its @Route URLs are an implementation detail and NOT part of that promise. A few areas are marked 'experimental' in their schema and may still change within the 1.x line.",
    "contact": {
      "name": "Peter Haller",
      "url": "https://github.com/phaller222/Prioritize",
      "email": "peter.haller@prioritize-iot.de"
    },
    "license": {
      "name": "Apache License 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.2.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "basicAuth": [],
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/tasks/{id}/goal/{goalId}": {
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Assign goal",
        "operationId": "task_assignGoal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Returns the projects the caller manages or participates in",
        "operationId": "project_getMyProjects",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Creates a new project (the caller becomes manager and first member)",
        "operationId": "project_createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/goals": {
      "get": {
        "tags": [
          "Project Goals"
        ],
        "summary": "Get goals",
        "operationId": "projectGoal_getGoals",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectGoalDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project Goals"
        ],
        "summary": "Create goal",
        "operationId": "projectGoal_createGoal",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectGoalDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/goals/{goalId}": {
      "get": {
        "tags": [
          "Project Goals"
        ],
        "summary": "Get goal",
        "operationId": "projectGoal_getGoal",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectGoalDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Project Goals"
        ],
        "summary": "Delete goal",
        "operationId": "projectGoal_deleteGoal",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Project Goals"
        ],
        "summary": "Update goal",
        "operationId": "projectGoal_updateGoal",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectGoalDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get project",
        "operationId": "project_getProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete project",
        "operationId": "project_deleteProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update project",
        "operationId": "project_updateProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/progress": {
      "get": {
        "tags": [
          "Project Goals"
        ],
        "summary": "Returns the project's progress derived from its goals and their tasks",
        "operationId": "projectGoal_getProgress",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectProgress"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks/{id}/goal": {
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Unassign goal",
        "operationId": "task_unassignGoal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TaskDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "CREATED",
              "ESTIMATED",
              "OPEN",
              "ASSIGNED",
              "STARTED",
              "STOPPED",
              "FINISHED",
              "CANCELLED",
              "CLOSED"
            ]
          },
          "assigneeId": {
            "type": "integer",
            "format": "int64"
          },
          "goalId": {
            "type": "integer",
            "format": "int64"
          },
          "processInstanceId": {
            "type": "string"
          },
          "tracking": {
            "type": "boolean"
          }
        }
      },
...
      "ProjectDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "beginDate": {
            "type": "string",
            "format": "date"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "maxManDays": {
            "type": "integer",
            "format": "int32"
          },
          "managerId": {
            "type": "integer",
            "format": "int64"
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "blackboardId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "beginDate": {
            "type": "string",
            "format": "date"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "maxManDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "GoalRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "properties": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ProjectGoalPropertyDocument"
                },
                {
                  "$ref": "#/components/schemas/ProjectGoalPropertyNumeric"
                }
              ]
            }
          }
        }
      },
      "ProjectGoalProperty": {
        "type": "object",
        "discriminator": {
          "propertyName": "type"
        },
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "ProjectGoalPropertyDocument": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProjectGoalProperty"
          },
          {
            "type": "object",
            "properties": {
              "tag": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ProjectGoalPropertyNumeric": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProjectGoalProperty"
          },
          {
            "type": "object",
            "properties": {
              "min": {
                "type": "number",
                "format": "double"
              },
              "max": {
                "type": "number",
                "format": "double"
              }
            }
          }
        ]
      },
      "ProjectGoalDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "GoalProgress": {
        "type": "object",
        "properties": {
          "goalId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "percentage": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProjectProgress": {
        "type": "object",
        "properties": {
          "overallPercentage": {
            "type": "integer",
            "format": "int32"
          },
          "goals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalProgress"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "HTTP Basic authentication against the local user store (default/dev profiles).",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "description": "Keycloak-issued JWT bearer token (keycloak profile).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}